change mqoi_desc_t head to be uint32_t so following fields are aligned when interpreting as uint32_t

+++ src/mini_qoi.h
@@ -61,7 +61,7 @@ typedef union {
 } mqoi_rgba_t;
 
 typedef struct {
-    uint8_t head;
+    uint32_t head;
 
     uint8_t magic[4];
     uint8_t width[4]; // big-endian width
